✨ Add Complete Register Form component
Added CompleteRegisterForm component for signing up with required fields: ID card, name, and class name. Included form validation using Zod schema. Updated CompleteSignUp and Profile components to utilize the new form for completing registration. Updated StudentClassesCard component to redirect to the profile page if required fields are missing.
