+++ online-assignment/components/_component/Profile.tsx
@@ -1,7 +1,9 @@
 "use client"
 import { useGetUserQuery } from "@/lib/features/userSlice";
-import ProfileHeader from "./ProfileHeader";
+import { Card } from "../ui/card";
+import CompleteRegisterForm from "./CompleteRegisterForm";
 import Loading from "./Loading";
+import ProfileHeader from "./ProfileHeader";
 
 type ProfileProps = {
   id:string;
@@ -10,10 +12,13 @@ const Profile = ({ id }:ProfileProps) => {
     const {data:user,isFetching} = useGetUserQuery(id)
   return (          
     isFetching ? <Loading/>: (
-    <div className="flex flex-col justify-center h-screen container mx-auto">
+    <div className="flex flex-col my-24 md:my-0 justify-center h-screen container mx-auto">
       <section>
       <ProfileHeader user={user!}/>
       </section>
+      <Card className="p-5 rounded-t-none shadow-sm">
+        <CompleteRegisterForm role={user?.role!} user={user!}/>
+      </Card>
     </div>
     )
   )
